WaitFor improvements:
added guard clauses.
usage of own gameobject and monobehaviour to make sure there are no sideeffects with other code.
cleanup for own gameobject and monobehaviour.
added a mandatory failure mode so that WaitFor is not going on forever and wasting resources if predicate is never satisfied.
changed parameterlist order so that stopPollingAfterInSec and pollRateInSec can have default values.
fixed minor formatting issues.